fixup! cpu/sam0_common: implement timer_set_periodic()

+++ cpu/sam0_common/periph/timer.c
@@ -237,7 +237,7 @@ int timer_set_absolute(tim_t tim, int channel, unsigned int value)
     return 0;
 }
 
-int timer_set_periodic(tim_t tim, int channel, unsigned int value, unsigned flags)
+int timer_set_periodic(tim_t tim, int channel, unsigned int value, uint8_t flags)
 {
     DEBUG("Setting timer %i channel %i to %i (repeating)\n", tim, channel, value);
